# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InstagramPostContentType = typeof InstagramPostContentType[keyof typeof InstagramPostContentType];


export const InstagramPostContentType = {
  Reel: 'Reel',
  Carousel: 'Carousel',
  Image: 'Image',
  Story: 'Story',
} as const;

export interface InstagramPost {
  id: string;
  date: string;
  caption: string;
  contentType: InstagramPostContentType;
  reach: number;
  likes: number;
  comments: number;
  saves: number;
  shares: number;
  engagementRate: number;
}

export interface InstagramDataInput {
  posts: InstagramPost[];
}

export type InstagramMetricTrend = typeof InstagramMetricTrend[keyof typeof InstagramMetricTrend];


export const InstagramMetricTrend = {
  up: 'up',
  down: 'down',
  neutral: 'neutral',
} as const;

export interface InstagramMetric {
  label: string;
  value: string;
  change: string;
  trend: InstagramMetricTrend;
}

export interface ContentTypePerformance {
  type: string;
  posts: number;
  reach: number;
  engagement: number;
  engagementRate: number;
}

export interface TrendPoint {
  label: string;
  reach: number;
  engagement: number;
}

export interface InstagramAnalysis {
  accountName: string;
  period: string;
  metrics: InstagramMetric[];
  posts: InstagramPost[];
  topPosts: InstagramPost[];
  underperformingPosts: InstagramPost[];
  contentTypes: ContentTypePerformance[];
  trends: TrendPoint[];
}

export interface AgentQuestionInput {
  /** @minLength 1 */
  question: string;
  analysis: InstagramAnalysis;
}

export interface AgentResponse {
  answer: string;
  tools: string[];
  citations: string[];
}

export interface RecommendationInput {
  analysis: InstagramAnalysis;
}

export type RecommendationPriority = typeof RecommendationPriority[keyof typeof RecommendationPriority];


export const RecommendationPriority = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface Recommendation {
  title: string;
  description: string;
  evidence: string;
  priority: RecommendationPriority;
  type: string;
}

export interface ContentPlanInput {
  analysis: InstagramAnalysis;
}

export interface ContentPlanDay {
  day: string;
  format: string;
  topic: string;
  hook: string;
  goal: string;
}