# lib/api-zod/src/generated/types/agentQuestionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InstagramAnalysis } from './instagramAnalysis';

export interface AgentQuestionInput {
  /** @minLength 1 */
  question: string;
  analysis: InstagramAnalysis;
}