# lib/api-zod/src/generated/types/contentPlanDay.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentPlanDay {
  day: string;
  format: string;
  topic: string;
  hook: string;
  goal: string;
}