# lib/api-zod/src/generated/types/contentTypePerformance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentTypePerformance {
  type: string;
  posts: number;
  reach: number;
  engagement: number;
  engagementRate: number;
}