# lib/api-zod/src/generated/types/instagramAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentTypePerformance } from './contentTypePerformance';
import type { InstagramMetric } from './instagramMetric';
import type { InstagramPost } from './instagramPost';
import type { TrendPoint } from './trendPoint';

export interface InstagramAnalysis {
  accountName: string;
  period: string;
  metrics: InstagramMetric[];
  posts: InstagramPost[];
  topPosts: InstagramPost[];
  underperformingPosts: InstagramPost[];
  contentTypes: ContentTypePerformance[];
  trends: TrendPoint[];
}