# lib/api-zod/src/generated/types/instagramDataInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InstagramPost } from './instagramPost';

export interface InstagramDataInput {
  posts: InstagramPost[];
}