# lib/api-zod/src/generated/types/instagramMetric.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InstagramMetricTrend } from './instagramMetricTrend';

export interface InstagramMetric {
  label: string;
  value: string;
  change: string;
  trend: InstagramMetricTrend;
}