# lib/api-zod/src/generated/types/instagramPost.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InstagramPostContentType } from './instagramPostContentType';

export interface InstagramPost {
  id: string;
  date: string;
  caption: string;
  contentType: InstagramPostContentType;
  reach: number;
  likes: number;
  comments: number;
  saves: number;
  shares: number;
  engagementRate: number;
}