# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationPriority } from './recommendationPriority';

export interface Recommendation {
  title: string;
  description: string;
  evidence: string;
  priority: RecommendationPriority;
  type: string;
}